Refactor, delete ROOM_CLOSE reason
