+++ mock/control-api/src/api/room.rs
